The Protocol part II
* framing for the response side of the protocol; structs
* minimal handling of framed responses on SW side, needs more work
* *_store_result() on test side of processing
* handling failure details for mem test

+++ fw/serial.c
@@ -72,7 +72,7 @@ void serial_rx_bytes(uint8_t *data, uint16_t count)
 }
 
 // -----------------------------------------------------------------------
-void serial_tx_bytes(uint8_t *data, uint8_t count)
+void serial_tx_bytes(uint8_t *data, uint16_t count)
 {
 	while (count--) {
 		serial_tx_char(*(data++));
